finished with constant pool parsing

+++ ClassFileParsing/Antlrv4BasedVersion/TestData.java
@@ -1,7 +1,7 @@
 public class TestData implements java.io.Serializable {
 
 	@Deprecated
-	final private int value = -2;
+	final private long value = -2;
 
 	public static void main(String[] args) {
 		for (int i = 0; i < 1; i++)
